tests/qsysfs: Add test to check if device entry exists.
Check if specified syspath entry exists. Normally when a device is
added a sysfs files are properly created, check the presence of one
of this files can be useful to check that a device is probed. Note,
but, some drivers create these files without probing or checking
the communication with the device, in that case this test does not
work and produces a false positive.

To avoid this always make sure that the sys entry is created only
when the device is probed, with some kind of communication between
the driver and the device, and is not created when the probe fails.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>