helpers/udaqlite: The uDAQliteServer class for our ammeter server.
The uDAQ-Lite is a low cost USB data acquisition device that incorporates A/D,
D/A, digital I/O and counter-timer functions. It supports 8 analog input
channels, 2 analog output channels, 8 input and 8 digital I/O and 1 counter
timer channel.

Here, the uDAQ-Lite, is used as a ammeter, using a shunt resitor to measure
the current.

Signed-off-by: Enric Balletbo i Serra <eballetbo@iseebcn.com>